apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
pypdf==5.9.0
pytest==8.4.1
httpx==0.28.1